+++ src/app/dashboard/dashboard.component.ts
@@ -1,6 +1,6 @@
 import { Component, OnInit } from '@angular/core';
 
-import { Project } from '../models/project';
+import { Project } from '../models/project.model';
 import { ProjectService } from '../services/project.service';
 
 @Component({
