+++ src/app/models/data-point.model.ts
@@ -1,0 +1,4 @@
+export class DataPoint {
+    date: Date;
+    value: number;
+}
