+++ src/app/models/indicator.model.ts
@@ -1,0 +1,8 @@
+export class Indicator {
+    name: string;
+    label: string;
+    description: string;
+    time_aggregation: string;
+    variables: string[];
+}
+
