+++ src/app/models/multi-data-point.model.ts
@@ -1,0 +1,8 @@
+export class MultiDataPoint {
+    date: Date;
+    values: {
+        'avg': number,
+        'min': number,
+        'max': number
+    };
+}
