+++ src/app/models/scenario.model.ts
