+++ src/app/project/project-form.ts
@@ -1,4 +1,4 @@
-import { Project } from '../models/project';
+import { Project } from '../models/project.model';
 
 export class ProjectForm {
