+++ src/app/services/scenario.service.ts
@@ -2,7 +2,7 @@ import { Injectable } from '@angular/core';
 import { Observable } from "rxjs";
 import 'rxjs/Rx';
 
-import { Scenario } from '../models/scenario';
+import { Scenario } from '../models/scenario.model';
 import { ApiHttp } from '../auth/api-http.service';
 import { apiHost } from "../constants";
